+++ src/main/java/com/iemr/hwc/service/common/transaction/CommonNurseServiceImpl.java
@@ -4165,10 +4165,10 @@ public class CommonNurseServiceImpl implements CommonNurseService {
 			for (Object[] objArr : benLastSixVisitDetails) {
 				vc = (String) objArr[1];
 				if (vc != null && vc.equalsIgnoreCase("Cancer Screening")) {
-					a = (BigInteger) objArr[2];
+					a = BigInteger.valueOf((long) objArr[2]);
 					benVisitCodeListCancer.add(a.longValue());
 				} else {
-					a = (BigInteger) objArr[2];
+					a = BigInteger.valueOf((long) objArr[2]);
 					benVisitCodeListOther.add(a.longValue());
 				}
 			}
